feat(m4): migrate canonical planning graph to workflows/planning.py
- Move build_pipeline and _node_policy from pipeline.py to
  arnold_pipelines/megaplan/workflows/planning.py as the authored source.
- Refactor pipeline.py into a thin facade that imports from workflows.planning.
- Update planning/operations.py to import build_pipeline from the canonical
  workflows.planning source.
- Add tests for workflow components and the authored planning workflow,
  including facade identity and manifest-hash-neutrality of prompt provenance.
- Preserve manifest/topology hashes and all existing public API contracts.
